+++ src/models/Settings.php
@@ -15,6 +15,8 @@ class Settings extends Model
     public $totpDelay = null;
 
     public $verifyFrontEnd = false;
+    public $verifyBackEnd = true;
+
     public $forceFrontEnd = false;
     public $forceBackEnd = false;
 
@@ -54,7 +56,7 @@ class Settings extends Model
     public function rules()
     {
         return [
-            [['verifyFrontEnd', 'forceFrontEnd', 'forceBackEnd'], 'boolean'],
+            [['verifyFrontEnd', 'verifyBackEnd', 'forceFrontEnd', 'forceBackEnd'], 'boolean'],
         ];
     }